Updated some deprecated library functions
Swapped some usages of TalonSRX from CTRE to PWMTalonSRX. This required some re-configuring on how encoders are configured since the pathfinder configureEncoder method only takes int, int, double parameters. The encoder position is now obtained using .getPosition() instead of .getSelectedSensorPosition(0). This requires rounding of the position (in order to provide the int parameter for the configureEncoder() method) and may result in loss in accuracy. Be sure to test this later