'Completed' OI Script
The OI script exists, and in theory should be functional, but in a loose sense. None of the commands that would be used by OI exist yet, so OI currently does nothing. The button command assignments used in Java are in the OI script commented out until commands are created.